lisa/tests: Rename test_bundle.py into base.py
