+++ lisa/tests/scheduler/eas_behaviour.py
@@ -28,7 +28,7 @@ from bart.common.Utils import area_under_curve
 from devlib.target import KernelVersion
 
 from lisa.wlgen.rta import Periodic, Ramp, Step
-from lisa.tests.test_bundle import ResultBundle, CannotCreateError, RTATestBundle
+from lisa.tests.base import ResultBundle, CannotCreateError, RTATestBundle
 from lisa.env import TestEnv
 from lisa.utils import ArtifactPath
 from lisa.energy_model import EnergyModel
