+++ lisa/tests/staging/load_tracking.py
@@ -18,7 +18,7 @@
 from bart.common.Utils import area_under_curve
 
 from lisa.wlgen.rta import RTATask, Periodic
-from lisa.tests.test_bundle import TestMetric, ResultBundle, CannotCreateError
+from lisa.tests.base import TestMetric, ResultBundle, CannotCreateError
 from lisa.tests.scheduler.load_tracking import (
     UTIL_AVG_CONVERGENCE_TIME_S,
     UTIL_SCALE,
